ドアを開けるために必要なカギに対応した南京錠の色を変える処理

+++ PliesonBreak/Assets/Scripts/PlayerBase.cs
@@ -218,7 +218,7 @@ public class PlayerBase : MonoBehaviour
     /// </summary>
     void PlayerHaveKey()
     {
-        if (HaveId == Door.NeedKeyID)
+        if (HaveId == (int)Door.NeedKeyID)
         {
             Debug.Log("ƒhƒA‚ªŠJ‚¢‚½");
             Door.DoorOpen(true);
